RDCC-958: moving validators to validator & impl packages, verifying error message in ExceptionMapperTests, added verify() to any applicable repository mocks
