+++ src/main/java/uk/gov/hmcts/reform/professionalapi/controller/request/validator/UserProfileUpdateRequestValidator.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.professionalapi.controller.request;
+package uk.gov.hmcts.reform.professionalapi.controller.request.validator;
 
 import uk.gov.hmcts.reform.professionalapi.domain.UserProfileUpdatedData;
