+++ src/main/java/uk/gov/hmcts/reform/professionalapi/util/RefDataUtil.java
@@ -106,7 +106,7 @@ public class RefDataUtil {
         try (Response response =  userProfileFeignClient.getUserProfileById(user.getUserIdentifier())) {
 
             Class clazz = response.status() > 300 ? ErrorResponse.class : GetUserProfileResponse.class;
-            ResponseEntity responseResponseEntity = JsonFeignResponseHelper.toResponseEntity(response, clazz);
+            ResponseEntity responseResponseEntity = JsonFeignResponseUtil.toResponseEntity(response, clazz);
 
             if (response.status() > 300) {
                 ErrorResponse userProfileErrorResponse = (ErrorResponse) responseResponseEntity.getBody();
@@ -130,7 +130,7 @@ public class RefDataUtil {
 
 
             Class clazz = response.status() > 300 ? ErrorResponse.class : ProfessionalUsersEntityResponse.class;
-            ResponseEntity responseResponseEntity = JsonFeignResponseHelper.toResponseEntity(response, clazz);
+            ResponseEntity responseResponseEntity = JsonFeignResponseUtil.toResponseEntity(response, clazz);
             if (response.status() < 300) {
 
                 modifiedOrgProfUserDetails = updateUserDetailsForActiveOrganisation(responseResponseEntity, activeOrganisationDetails);
@@ -284,7 +284,7 @@ public class RefDataUtil {
         ModifyUserRolesResponse modifyUserRolesResponse = new ModifyUserRolesResponse();
         boolean isFailureFromUp = response.status() > 300;
         Class clazz = isFailureFromUp ? ErrorResponse.class : ModifyUserRolesResponse.class;
-        ResponseEntity responseEntity = JsonFeignResponseHelper.toResponseEntity(response, clazz);
+        ResponseEntity responseEntity = JsonFeignResponseUtil.toResponseEntity(response, clazz);
         if (isFailureFromUp) {
             ErrorResponse errorResponse = (ErrorResponse) responseEntity.getBody();
             modifyUserRolesResponse.setErrorResponse(errorResponse);
@@ -300,7 +300,7 @@ public class RefDataUtil {
         try (Response response =  userProfileFeignClient.getUserProfileByEmail(emailAddress)) {
 
             Class clazz = response.status() > 300 ? ErrorResponse.class : NewUserResponse.class;
-            ResponseEntity responseResponseEntity = JsonFeignResponseHelper.toResponseEntity(response, clazz);
+            ResponseEntity responseResponseEntity = JsonFeignResponseUtil.toResponseEntity(response, clazz);
 
             if (response.status() == 200) {
