+++ src/test/java/uk/gov/hmcts/reform/professionalapi/controller/request/validator/UpdateOrganisationRequestValidatorTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.professionalapi.controller.request;
+package uk.gov.hmcts.reform.professionalapi.controller.request.validator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
@@ -7,8 +7,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.junit.Test;
-import uk.gov.hmcts.reform.professionalapi.controller.request.OrganisationIdentifierValidator;
-import uk.gov.hmcts.reform.professionalapi.controller.request.UpdateOrganisationRequestValidator;
+import uk.gov.hmcts.reform.professionalapi.controller.request.validator.OrganisationIdentifierValidator;
+import uk.gov.hmcts.reform.professionalapi.controller.request.validator.UpdateOrganisationRequestValidator;
 
 public class UpdateOrganisationRequestValidatorTest {
